requirements.txt
Flask
Flask-Session
Flask-Cors
python-dotenv
PyPDF2
supabase
openai
stripe
requests
supabase
ics
langchain
langchain-experimental
langchain_community
openai
pymysql
psycopg2-binary
python-socketio